% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_deseq2_getDegs.R
\name{mrf_deseq2_getDegs}
\alias{mrf_deseq2_getDegs}
\title{Extract Differentially Expressed Genes}
\usage{
mrf_deseq2_getDegs(x, logFC = 1, pvalue = 0.05)
}
\arguments{
\item{x}{A DESeq2 results object containing the results of differential expression analysis.}

\item{logFC}{Numeric, the log2 fold change threshold for significance. Default is 1.}

\item{pvalue}{Numeric, the p-value threshold for significance. Default is 0.05.}
}
\value{
A character vector containing the symbols of differentially expressed genes (DEGs) that meet the specified thresholds.
}
\description{
This function extracts differentially expressed genes (DEGs) from a DESeq2 results object based on specified log fold change
and p-value thresholds.
}
\examples{
\dontrun{
library(DESeq2)
library(dplyr)
library(tidyr)
library(tibble)
# Example DESeq2 analysis
dds <- DESeqDataSetFromMatrix(countData = countData, colData = colData, design = ~ condition)
dds <- DESeq(dds)
res <- results(dds)
# Extract DEGs
degs <- mrf_deseq2_getDegs(res, logFC = 1, pvalue = 0.05)
print(degs)
}

}
