% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_deseq2_getTab.R
\name{mrf_deseq2_getTab}
\alias{mrf_deseq2_getTab}
\title{Extract Differential Expression Results into a Data Frame}
\usage{
mrf_deseq2_getTab(x)
}
\arguments{
\item{x}{A DESeq2 results object from which to extract the data. This object should contain list data with components
`baseMean`, `log2FoldChange`, `lfcSE`, `stat`, `pvalue`, and `padj`, as well as row names.}
}
\value{
A data frame with the following columns:
\item{baseMean}{Mean of normalized counts for all samples.}
\item{logFC}{Log2 fold change between conditions.}
\item{lfcSE}{Standard error of the log2 fold change estimate.}
\item{stat}{Test statistic value.}
\item{pvalue}{Raw p-value.}
\item{padj}{Adjusted p-value.}
}
\description{
This function extracts relevant data from a DESeq2 results object and returns it as a data frame.
}
\examples{
\dontrun{
library(DESeq2)
# Example DESeq2 analysis
dds <- DESeqDataSetFromMatrix(countData = countData, colData = colData, design = ~ condition)
dds <- DESeq(dds)
res <- results(dds)
# Extract results into a data frame
res_df <- mrf_deseq2_getTab(res)
head(res_df)
}

}
