% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_ont_pecat_cfgfile.R
\name{mrf_ont_pecat_cfgfile}
\alias{mrf_ont_pecat_cfgfile}
\title{Create a PECAT Configuration File for Oxford Nanopore Sequencing}
\usage{
mrf_ont_pecat_cfgfile(
  project = NULL,
  bed = NULL,
  reads = NULL,
  bam = NULL,
  genome_size = NULL,
  threads = 4,
  mode = c("Single", "Multiple")
)
}
\arguments{
\item{project}{`character`. Name of the project. Required in `Single` mode. Ignored in `Multiple` mode.}

\item{bed}{`character`. Path to the BED file specifying regions of interest. Required in `Multiple` mode.}

\item{reads}{`character`. Path to the reads file. Required in `Single` mode.}

\item{bam}{`character`. Path to the BAM file. Required in `Multiple` mode.}

\item{genome_size}{`numeric`. The size of the genome. Required in `Single` mode.}

\item{threads}{`integer`. Number of threads to use for processing. Default is 4.}

\item{mode}{`character`. Mode of operation. Can be either `"Single"` or `"Multiple"`. Default is `"Single"`.}
}
\value{
The function creates one or more configuration files based on the provided input. No return value.
}
\description{
This function generates a configuration file for the PECAT pipeline, tailored to Oxford Nanopore sequencing data. The function supports two modes of operation: `Single` and `Multiple`.
}
\details{
In `Single` mode, the function generates a configuration file for a single sequencing project based on the provided project name, reads, and genome size. In `Multiple` mode, the function processes a BED file with multiple regions and generates individual configuration files for each region using a BAM file.


- In `Single` mode, a single configuration file is generated with the specified project name, reads file, and genome size.
- In `Multiple` mode, the function reads a BED file and generates a configuration file for each region specified in the BED file using the corresponding region from the BAM file.
}
\note{
The function requires that the PECAT tool (`pecat.pl`) is available and that the user is in a Conda environment. If not, the function will stop with an error.
}
\examples{
\dontrun{
# Single mode example
mrf_ont_pecat_cfgfile(project = "my_project",
                      reads = "reads.fastq",
                      genome_size = 3000000000,
                      mode = "Single")

# Multiple mode example
mrf_ont_pecat_cfgfile(bed = "regions.bed",
                      bam = "alignment.bam",
                      mode = "Multiple")
}

}
