% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_ont_pecat_run.R
\name{mrf_ont_pecat_run}
\alias{mrf_ont_pecat_run}
\title{Run PECAT for Multiple Configuration Files}
\usage{
mrf_ont_pecat_run()
}
\value{
The function does not return a value but runs PECAT on all detected configuration files and provides console output on the progress.
}
\description{
This function automates the execution of the PECAT pipeline for multiple configuration files generated using the `mrf_ont_pecat_cfgfile` function. The function identifies all configuration files in the working directory with the `_cfgfile` suffix and runs the PECAT pipeline for each of them sequentially.
}
\details{
The function searches the current working directory for files with the `_cfgfile` suffix, then iteratively runs the `pecat.pl unzip` command on each configuration file. The progress is displayed in the console with the file's associated gene name highlighted.
}
\examples{
\dontrun{
# Run PECAT for all configuration files in the directory
mrf_ont_pecat_run()
}

}
