% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_primerDesign_mirnaMat.R
\name{mrf_primerDesign_mirnaMat}
\alias{mrf_primerDesign_mirnaMat}
\title{Design Primers for miRNA Mature}
\usage{
mrf_primerDesign_mirnaMat(x, print = FALSE)
}
\arguments{
\item{x}{A character string representing a miRNA sequence.}

\item{print}{A logical value indicating whether to print the primers to the console. Default is `FALSE`.}
}
\value{
If `print = FALSE`, the function returns a named vector containing the forward and reverse primers. If `print = TRUE`, the primers are printed to the console with color formatting, and nothing is returned.
}
\description{
This function designs forward and reverse primers for a given miRNA sequence using predefined sequences. The reverse primer is the complement of the latter part of the transcribed sequence.
}
\examples{
\dontrun{
# Example miRNA sequence
mirna_seq <- "UGAGGUAGUAGGUUGUAUAGUU"

# Generate primers and print them
mrf_primerDesign_mirnaMat(mirna_seq, print = TRUE)

# Generate primers and return them as a vector
primers <- mrf_primerDesign_mirnaMat(mirna_seq)
print(primers)
}

}
