% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_samtools_bam2fq.R
\name{mrf_samtools_bam2fq}
\alias{mrf_samtools_bam2fq}
\title{Convert BAM to FASTQ Using SAMtools}
\usage{
mrf_samtools_bam2fq(bam = NULL, output_name = NULL, threads = 1, region = NULL)
}
\arguments{
\item{bam}{`character`. Path to the input BAM file. This argument is required.}

\item{output_name}{`character`. Name of the output FASTQ file. If not specified, the output file will be named based on the input BAM file with a `_reads.fastq` suffix.}

\item{threads}{`integer`. Number of threads to use for the conversion process. Default is 1.}

\item{region}{`character`. Genomic region to extract in the format `chr:start-end`. If not specified, the entire BAM file is converted.}
}
\value{
`character`. The path to the generated FASTQ file.
}
\description{
This function converts a BAM file to a FASTQ file using SAMtools. It optionally allows for region-specific extraction and supports multi-threading.
}
\details{
The function first uses `samtools view` to extract the specified region from the BAM file (if a region is specified) and writes it to a temporary BAM file. Then, it converts the temporary BAM file to FASTQ format using `samtools bam2fq`. The temporary BAM file is deleted after conversion.
}
\note{
Ensure that SAMtools is installed and available in your system's PATH before running this function.
}
\examples{
\dontrun{
# Convert an entire BAM file to FASTQ
mrf_samtools_bam2fq(bam = "alignment.bam", output_name = "output_reads.fastq")

# Convert a specific region from the BAM file to FASTQ
mrf_samtools_bam2fq(bam = "alignment.bam", region = "chr1:1000-5000")
}

}
