% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_samtools_view.R
\name{mrf_samtools_view}
\alias{mrf_samtools_view}
\title{Extract and Convert Regions from BAM Files Using SAMtools}
\usage{
mrf_samtools_view(
  bam = NULL,
  output_format = "BAM",
  region = NULL,
  threads = 1,
  output_name = NULL
)
}
\arguments{
\item{bam}{`character`. Path to the input BAM file. This argument is required.}

\item{output_format}{`character`. The format of the output file. Options are `"BAM"`, `"SAM"`, or `"CRAM"`. Default is `"BAM"`.}

\item{region}{`character`. Genomic region to extract in the format `chr:start-end`. If not specified, the entire BAM file is processed.}

\item{threads}{`integer`. Number of threads to use for the operation. Default is 1.}

\item{output_name}{`character`. Name of the output file. If not specified, a default name is generated based on the input BAM file and the chosen output format.}
}
\value{
`character`. The path to the generated output file.
}
\description{
This function uses SAMtools to extract specified regions from a BAM file and convert the output to a chosen format (BAM, SAM, or CRAM). It supports multi-threading and allows users to specify an output file name.
}
\details{
The function constructs a SAMtools `view` command to extract data from the specified BAM file. The output can be saved in BAM, SAM, or CRAM format, depending on the user's choice. If a region is specified, only that region is extracted; otherwise, the entire BAM file is processed.
}
\note{
Ensure that SAMtools is installed and available in your system's PATH before running this function.
}
\examples{
\dontrun{
# Extract and save a region in BAM format
mrf_samtools_view(bam = "alignment.bam", region = "chr1:1000-5000")

# Convert the entire BAM file to SAM format
mrf_samtools_view(bam = "alignment.bam", output_format = "SAM")

# Extract and save a region in CRAM format
mrf_samtools_view(bam = "alignment.bam", region = "chr2:2000-6000", output_format = "CRAM")
}

}
