% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf_viz_vennDiagram.R
\name{mrf_viz_vennDiagram}
\alias{mrf_viz_vennDiagram}
\title{Create and Save a Venn Diagram}
\usage{
mrf_viz_vennDiagram(
  x,
  filename = "",
  res = 600,
  height = 6000,
  width = 6000,
  compression = "lzw",
  cex = 2,
  fontface = "bold",
  fill = ""
)
}
\arguments{
\item{x}{A list of sets for which the Venn diagram is to be created. The length of the list should be between 2 and 4.}

\item{filename}{A character string specifying the name of the output file. Default is an empty string.}

\item{res}{Numeric, the resolution of the output image. Default is 600.}

\item{height}{Numeric, the height of the output image. Default is 6000.}

\item{width}{Numeric, the width of the output image. Default is 6000.}

\item{compression}{A character string specifying the type of compression for the output file. Default is "lzw".}

\item{cex}{Numeric, the scaling factor for the text in the diagram. Default is 2.}

\item{fontface}{A character string specifying the font face for the text. Default is "bold".}

\item{fill}{A character vector specifying the fill colors for the sets. Default is an empty string.}
}
\value{
A Venn diagram saved to the specified file.
}
\description{
This function creates a Venn diagram based on the provided list of sets and saves it to a file.
}
\examples{
\dontrun{
library(VennDiagram)
# Example data
list_of_sets <- list(A = 1:5, B = 3:7, C = 6:10)
# Create and save Venn diagram
mrf_viz_vennDiagram(list_of_sets, filename = "venn_diagram.tiff")
}

}
