% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_seqComplement.R
\name{util_seqComplement}
\alias{util_seqComplement}
\title{Get the Complement of a DNA Sequence}
\usage{
util_seqComplement(x)
}
\arguments{
\item{x}{A character string representing a DNA sequence (containing the nucleotides A, T, C, G).}
}
\value{
A character string representing the complementary DNA sequence.
}
\description{
This function takes a DNA sequence as input and returns its complementary sequence.
}
\examples{
\dontrun{
dna_sequence <- "ATCG"
complement_sequence <- util_seqComplement(dna_sequence)
print(complement_sequence)  # Output will be "TAGC"
}

}
