% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_transcribe.R
\name{util_transcribe}
\alias{util_transcribe}
\title{Transcribe RNA Sequence to DNA Sequence}
\usage{
util_transcribe(x)
}
\arguments{
\item{x}{A character string representing an RNA sequence.}
}
\value{
A character string representing the corresponding DNA sequence.
}
\description{
This function transcribes an RNA sequence by replacing all occurrences of "U" with "T", converting it to a DNA sequence.
}
\examples{
\dontrun{
rna_sequence <- "AUGCUA"
dna_sequence <- util_transcribe(rna_sequence)
print(dna_sequence)  # Output will be "ATGCTA"
}

}
